+++ test/acceptance/tests/basic/terraform/basic-install/main.tf
@@ -274,19 +274,26 @@ module "test_server" {
   port = 9090
 
   tls                            = var.secure
-  consul_server_ca_cert_arn      = module.consul_server.ca_cert_arn
+  consul_server_ca_cert_arn      = var.secure ? module.consul_server.ca_cert_arn : ""
   gossip_key_secret_arn          = var.secure ? aws_secretsmanager_secret.gossip_key[0].arn : ""
   acls                           = var.secure
   consul_client_token_secret_arn = var.secure ? module.acl_controller[0].client_token_secret_arn : ""
   acl_secret_name_prefix         = var.suffix
   consul_ecs_image               = var.consul_ecs_image
 
-  additional_task_role_policies = [aws_iam_policy.execute-command.arn]
+  // Test passing in roles. This requires users to correctly configure the roles outside mesh-task.
+  create_task_role      = false
+  create_execution_role = false
+  task_role             = aws_iam_role.task
+  execution_role        = aws_iam_role.execution
+
 }
 
-// Testing passing task/execution role into mesh-task
+// Configure a task role for passing in to mesh-task.
 resource "aws_iam_role" "task" {
   name = "test_server_${var.suffix}_task_role"
+  path = "/consul-ecs/"
+
   assume_role_policy = jsonencode({
     Version = "2012-10-17"
     Statement = [
@@ -301,6 +308,35 @@ resource "aws_iam_role" "task" {
   })
 }
 
+// Policy to allow iam:GetRole (required for auth method)
+resource "aws_iam_policy" "get-task-role" {
+  count = var.secure ? 1 : 0
+  name  = "test_server_${var.suffix}_get_task_role_policy"
+  path  = "/consul-ecs/"
+
+  policy = <<EOF
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Effect": "Allow",
+      "Action": [
+        "iam:GetRole"
+      ],
+      "Resource": [
+        "${aws_iam_role.task.arn}"
+      ]
+    }
+  ]
+}
+EOF
+}
+
+resource "aws_iam_role_policy_attachment" "get-task-role" {
+  count      = var.secure ? 1 : 0
+  role       = aws_iam_role.task.id
+  policy_arn = aws_iam_policy.get-task-role[count.index].arn
+}
 
 // Policy to allow `aws execute-command`
 resource "aws_iam_policy" "execute-command" {
@@ -330,6 +366,11 @@ EOF
   # tags = var.tags
 }
 
+resource "aws_iam_role_policy_attachment" "execute-command" {
+  role       = aws_iam_role.task.id
+  policy_arn = aws_iam_policy.execute-command.arn
+}
+
 resource "aws_iam_role" "execution" {
   name = "test_server_${var.suffix}_execution_role"
   path = "/ecs/"
@@ -346,6 +387,37 @@ resource "aws_iam_role" "execution" {
       }
     ]
   })
+
+  inline_policy {
+    name   = "test_server_${var.suffix}_execution_role_policy"
+    policy = <<EOF
+{
+  "Version": "2012-10-17",
+  "Statement": [
+%{if var.secure~}
+    {
+      "Effect": "Allow",
+      "Action": [
+        "secretsmanager:GetSecretValue"
+      ],
+      "Resource": [
+        "${module.consul_server.ca_cert_arn}",
+        "${aws_secretsmanager_secret.gossip_key[0].arn}"
+      ]
+    },
+%{endif~}
+    {
+      "Effect": "Allow",
+      "Action": [
+        "logs:CreateLogStream",
+        "logs:PutLogEvents"
+      ],
+      "Resource": "*"
+    }
+  ]
+}
+EOF
+  }
 }
 
 locals {